Ignore trailing slashes in History location API
For getURL, setURL, and formatURL, always strip trailing slashes from
url.
Set the default rootURL back to the empty string.
Add support for baseURL and rootURL with and without
leading and trailing slashes.
Fix the way the baseURL is being set in the test to match
style more closely.
Fix broken overriding of the pushState call in the tests.
Fix broken test that is missing the rootURL in the initial state.